extract structs from api_server.rs to its own files
