+++ rust/sglang-server/src/api_server/submit.rs
@@ -5,7 +5,8 @@
 use axum::{http::StatusCode, response::Response};
 use tokio::sync::mpsc;
 
-use super::{AppState, native_api::native_error};
+use super::app::AppState;
+use super::native_api::native_error;
 use crate::message::egress::{EgressItem, EgressSink};
 use crate::message::ids::Rid;
 use crate::message::request::{Request, RequestKind};
